requirements.txt
Django==4.1.6
django-bootstrap5==22.2
django_debug_toolbar==3.8.1
psycopg2-binary==2.9.5
Pillow==9.4.0
djangorestframework==3.14.0
